+++ lib/generator_helper.dart
@@ -1,8 +1,8 @@
 library prisma.generator_helper;
 
-export 'generator_helper/src/binary_targets_env_value.dart';
-export 'generator_helper/src/config.dart';
-export 'generator_helper/src/env_value.dart';
-export 'generator_helper/src/generator.dart';
-export 'generator_helper/src/manifest.dart';
-export 'generator_helper/src/options.dart';
+export 'src/generator_helper/binary_targets_env_value.dart';
+export 'src/generator_helper/config.dart';
+export 'src/generator_helper/env_value.dart';
+export 'src/generator_helper/generator.dart';
+export 'src/generator_helper/manifest.dart';
+export 'src/generator_helper/options.dart';