+++ lib/src/generator_helper/options.dart
@@ -1,4 +1,4 @@
-import '../../dmmf/src/dmmf.dart';
+import '../dmmf/dmmf.dart';
 import 'config.dart';
 import 'env_value.dart';